cloud_functions/gemini-pro-news-custom/requirements.txt
functions-framework==3.*
gnews
requests
google-generativeai
# google-cloud-aiplatform